11/08 09:03 끝

+++ app/api/service/photo_service.py
@@ -12,15 +12,18 @@ from app import s3
 # 포토 업로드
 from app.api.repository.album_repository import album_repository
 from app.api.repository.photo_repository import photo_repository
+from app.api.repository.user_repository import user_repository
 
 from app.api.model.photo import Photo
 
-from app.api.custom_exception.common_exception import ForbiddenException
+from app.api.custom_exception.common_exception import ForbiddenException, NotExistPhoto, NotExistAlbum
 
 
 class PhotoService:
 
+    # 사진 업로드
     def photo_upload(self, user_idx: int, file: werkzeug.datastructures.FileStorage, album_idx: int):
+
         album = album_repository.find_album_by_album_idx(album_idx=album_idx)
 
         # album_idx 를 해당 User 가 가지고 있는지 확인
@@ -42,8 +45,9 @@ class PhotoService:
 
         return True
 
-    # 포토 다운로드
+    # 사진 다운로드
     def photo_download(self, key: str):
+
         data = s3.get_object(Bucket='knight2995-photo-album', Key=key).get('Body').read()
         img = cv2.imdecode(np.fromstring(data, np.uint8), cv2.IMREAD_UNCHANGED)
 
@@ -52,15 +56,27 @@ class PhotoService:
         return json.dumps({"imgData": base64.b64encode(buffer).decode('utf-8')})
 
     # 모든 사진 조회(앨범 내의)
-    def find_all_photos(self, album_idx: int):
+    def find_all_photos(self, album_idx: int, user_idx: int):
+
+        # 앨범이 존재하는 지 여부 확인
+        album = album_repository.find_album_by_album_idx(album_idx)
+
+        if album is None:
+            raise NotExistAlbum
+
+        # album_idx 를 User 가 가지고 있는지 확인
+        if album.user_idx != user_idx:
+            raise ForbiddenException
+
         photos = photo_repository.find_photos_by_album_idx(album_idx)
 
         # 직접 변환
         photos_data = list(map(lambda photo: {'idx': photo.idx, 'image_key': photo.image_key}, photos))
         return photos_data
 
-    # idx에 해당하는 사진 조회 후 반환
+    # idx 에 해당하는 사진 조회 후 이미지 데이터 반환
     def find_photo_data(self, idx: int):
+
         photo = photo_repository.find_photo_by_idx(photo_idx=idx)
 
         data = s3.get_object(Bucket='knight2995-photo-album', Key=photo.image_key).get('Body').read()
@@ -70,17 +86,35 @@ class PhotoService:
 
         return base64.b64encode(buffer).decode('utf-8')
 
-    def delete_photo(self, photo_idx: int):
+    # s3에서 실제 이미지 삭제
+    def delete_photo(self, image_key):
 
-        """ Todo 검증 코드 필요함 """
-        self.delete_photo_data([photo_repository.find_photo_by_idx(photo_idx)])
+        s3.delete_object(Bucket='knight2995-photo-album', Key=image_key)
+
+    # photo_idx 에 해당하는 이미지 삭제
+    def delete_photo_by_idx(self, photo_idx: int, user_idx: int):
+
+        # 사진의 존재 여부 확인
+        photo = photo_repository.find_photo_by_idx(photo_idx)
+
+        if photo is None:
+            raise NotExistPhoto
+
+        # 현재 user 가 가진 앨범 안에 존재하는 사진 여부 확인
+        albums = album_repository.find_albums_by_user_idx(user_idx)
+        album_idx_list = list(map(lambda album: album.idx, albums))
+
+        if photo.album_idx not in album_idx_list:
+            raise ForbiddenException
+
+        self.delete_photo(photo_repository.find_photo_by_idx(photo_idx).image_key)
         photo_repository.delete_photo_by_idx(photo_idx)
 
-    # s3에서 삭제
-    def delete_photo_data(self, photos: list):
+    # 사진 리스트 단위 삭제
+    def delete_photos_data(self, photos: list):
 
         for photo in photos:
-            s3.delete_object(Bucket='knight2995-photo-album', Key=photo.image_key)
+            self.delete_photo(photo.image_key)
 
 
 photo_service = PhotoService()
